requirements.txt
Flask==3.0.3
flask-mysqldb==2.0.0
mysqlclient==2.2.4
Werkzeug==3.0.3
